[Tf] Implement TfDebug::SetOutputFile in python 3
For simplicity in the code I've changed the exception type that is
raised when passing a parameter of the wrong type. It was TypeError,
and will now be Tf.ErrorException. The only valid arguments to this
function are stderr and stdout, so I think this is more clear, since
the Tf.ErrorException will indicate that. It also lets me avoid
potentially complicated code in python 3's C implentation to determine
whether the passed in type is a "file like" object.

This was simpler in python 2, because python files were thin wrappers
on the underlying file system handles. That's not true any more in
python 3.