[Tf] Rename and move compatibility utils into their own header
The new symbols used for python version compatibility will now have
a Tf prefix, and are in their own header file. I also removed some
unused symbols, and simplified the PyString_AsStdString function.

There is a python 3 equivalent to PyString_AsString called
PyUnicode_AsUTF8. It is the recommended path forward for writing C
libraries that support both versions of python, so I switched to
using it. This also allows us to only convert to std::string when
needed.