[usdviewq] Remove deprecated settings class
This is more of a proposal than a real change. This appears to be
completely unused. If I make this change and build all tests pass,
and usdview seems to work fine. Deleting settings.py would let us avoid
writing backwards compatibility code for its file format. This wouldn't
be a ton of work, but I don't want to do it if this file is completely
vestigial. I vote we amputate it.

If this is something we can agree to, we can have a followup change that
renames settings2 to settings, and just have the json settings file from
now on, instead of maintaining two files (one of which appears to be
mostly empty and completely unused).