[gf] Support for "future division" in python 2
Division works correctly in python 2 and python 3, but if you use
"from __future__ import division" in python 2, the custom division
operators defined in Gf were not being invoked. This is because they
call into __truediv__ instead of __div__. In python 3 builds boost
handles that for you, but it doesn't do so in python 2.