Enable builds for distribution on PyPI (for pip)
This requires a few tweaks to the build parameters, but most of these
can be accomplished on the command line for build_usd.py. A new
parameter was added to enable a few changes to the build that aren't
possible through the command line.

Builds are expected to be run on azure-pipelines, so this build process
is set up through a new pipeline file.