Experimental refactoring of Python module initialization
This change adds a new Tf.PreparePythonModule function that
is intended to replace the old Tf.PrepareModule function.
The new function takes care of importing the underlying
boost::python module (i.e. "_<libname>"), which allows us
to remove a small amount of boilerplate from all of the
__init__.py files. It gives us a single place where we can
use the new Tf.WindowsImportWrapper rather than scattering
it throughout the entire codebase. We could also use this
to do other Python initialization if needed in the future.