usdShaders: Move usdShaders to pxr/usd/plugin
This makes the "shipped" shaders like UsdPreviewSurface available
for queries via Sdr in OpenUSD builds with the imaging component
disabled.

In particular, this allows usdchecker to validate assets that use
these shaders in no-imaging builds. For example, usdchecker would
previously fail in these builds when it encountered a Shader prim
using UsdPreviewSurface because that shader and its definition file
would not be installed.

Note that the .glslfx and .osl implementations for these shaders have
also been moved to keep them next to the corresponding .usda definition
file. However, the code that actually parses the .glslfx files still
lives in pxr/imaging. This assumes that clients that want to consume
the shader implementation will need to have imaging enabled.

The sdrGlslfx parser plugin that populates Sdr directly from .glslfx
files also remains in pxr/imaging. This isn't needed for these shaders
since their definitions are read from shaderDefs.usda. We may need
to move this in the future if there are other shipped .glslfx shaders
without corresponding .usda definitions that need to be accessible
via Sdr.

Fixes #3055

(Internal change: 2358070)