[hdx] Add mutex to protect concurrent access to shared `HdSelection`.
This commit adds a mutex that the caller is responsible for locking when it is about to concurrently mutate the shared `HdSelection` object.

This fixes a race condition that was a result of `_PopulateFromSceneIndex` concurrently invoking `_AddToSelection`.  `_AddToSelection` calls the mutable functions `AddInstance` or `AddRprim`, neither of which are safe to call concurrently since they mutate a `std::unordered_map`.

(Internal change: 2358165)