This change alters the behavior of SdfAssetPath in a few ways:
Renames the existing _assetPath member and meaning of first parameter to the constructor to now represent the raw authored path in the layerr.  Added explicit getters for this authored path. It is important to note that when authoring scene description, this value will be used for serilization, all other fields will be ignored.

Adds a new evaluatedPath field to SdfAssetPath.  This field is meant to hold the original asset path as authored in the layer with any variable expressions evaluated.  This field is only populated if passed in at construction time.

Changes the behavior of SdfAssetPath::GetAssetPath to return the evaulated path if present, otherwise returns the raw authored path.  This is to preserve compatibility with existing callers who were relying on this method to return them the unresolved asset path containing evaulated expression variables.

Additionally, _MakeResolvedAssetPathsImpl in UsdStage has been augmented to provide the unresolved, evaluated path when creating SdfAssetPath objects.  This function services calls to UsdAttribute::Get().

(Internal change: 2358824)