sdf: Fix a bug where we were not correctly handling path patterns with "bare predicates" following stretch, like `//{pred}...`.
For example, the pattern `/foo//{pred}/bar` should match the path
`/foo/bar` if `/foo` passes {pred}.  The way that matching proceeds is
to split patterns into components by //, and match the components.  The
previous code assumed that components could not overlap, which is true
*except* in just this one case.  So now for segments that start with
bare predicates (i.e. //{pred}/...) we look for matches that overlap by
one with the prior component.

(Internal change: 2358827)