build_usd.py: Apply fix to OpenVDB to support Xcode 16.3
AcademySoftwareFoundation/openvdb#1977 fixed a syntax issue that
caused errors when building with Xcode 16.3. This change updates
build_usd.py to apply the (simple) patch to the version of OpenVDB
it builds. We manually back-port this fix for a few reasons:

- The PR was merged in December 2024, but is currently not part
  of an official release.

- build_usd.py currently uses OpenVDB 9.1.0, which is several
  major versions behind the current latest version 12.0.0.
  Even if the PR were released in a new 12.x version, we wouldn't
  want to jump all the way there since it's so far ahead of what
  we're building and testing against internally.

If OpenVDB officially ports the fix back to older versions, we
can revert this change and just update build_usd.py to pull in
the official release instead.

(Internal change: 2359154)