Autodesk: testUsdImagingGL image diff threshold adjustments
Slightly increase `testUsdImagingGL` image diff thresholds to allow them to
pass with Metal and Vulkan (Lavapipe).

For tests for which it was relevant to do so, these threshold increases were
copied to the studio test versions.

The threshold increase to testUsdImagingGLBasisCurvesWithColor in the PR was
removed, as I believe the image diff is significant enough to warrant further
investigation.

Closes #3461

(Internal change: 2359846)