pcp: Add test case for bug with specializes and ancestral arcs
This test case was taken from testUsdInherits, with some modifications.
It demonstrates a bug where, in certain cases, specializes nodes from
recursive prim index computations for ancestral arcs are not propagated
to the root, which causes their opinions to have the wrong strength
order.

The /Parent/Child prim in this test runs into this bug -- ancestral
specializes opinions from /PIS/Sibling and /PISI/Sibling are in the
wrong strength order. /Parent2/Child is set up the same way as
/Parent/Child, but with an additional specializes arc authored on
/Parent2. This additional arc happens to work around the bug, giving
this prim the expected strength ordering.

(Internal change: 2359950)