pcp: Fix incorrect strength order for ancestral specializes opinions
This change fixes the bug described in change 2359950.
Prim indexes for ancestral opinions are now fully scanned to
determine if there are any specializes nodes that need to be
propagated for strength ordering purposes.

Previously, we only scanned the first level of child nodes in
that prim index. That assumed that specializes nodes would have
been propagated to that level as part of the recursive prim
index computation. That is true for inherits nodes, but not for
specializes.

The SpecializesAndAncestralArcs4 baseline has been updated to
show the correct behavior: the opinions from the ancestral
specializes nodes in /Parent/Child now match the strength order
in /Parent2/Child, which did not exhibit the original problem.

(Internal change: 2360374)