pcp: Add test for bug with "is due to ancestor" flag and specializes
This change adds a test case to the SpecializesAndAncestralArcs4 museum
example that dumps detailed information for the /Parent/Child prim.

The propagated specializes subtree at nodes 10 and 11 are marked as
*not* due to ancestors, which is incorrect. The corresponding source
nodes 3 and 4 are marked as due to ancestor, which should have been
copied over to nodes 10 and 11 as part of the propagation process.
This will be fixed in an upcoming change.

(Internal change: 2360750)