pcp: Copy "is due to ancestor" bit when propagating specializes
This fixes the bug noted in change 2360750, where nodes in propagated
specializes subtrees would be incorrectly marked as non-ancestral.
The baseline update for SpecializesAndAncestralArcs4 shows the effect
of this fix: nodes 10 and 11 now report the same "is due to ancestor"
as their source nodes 3 and 4.

This issue showed up in testUsdInherits; in test_GetAllDirectInherits,
the call to UsdInherits::GetAllDirectInherits for /Parent/Child returned
a few paths that, on closer inspection, it shouldn't have. This test
has been fixed up accordingly.

Fixing this bug revealed a separate issue: Pcp had been relying on the
old behavior to ensure the nodes in the specializes subtree would be
included in Pcp's dependency tracking. For example, in the test case
above, the fix caused nodes 10 and 11 to be classified as "ancestral"
dependencies instead of "purely direct". Since Pcp does not track
ancestral dependencies, edits to these nodes' prim specs would not
invoke change processing as expected. This showed up in the namespace
editing tests in pxr/usd/usd.

PcpClassifyNodeDependency was updated to traverse to the origin of any
propagated specializes nodes it encounters while walking from the given
node to the root of the prim index. This ensures that dependency
information is captured from the site where the specializes arc was
originally introduced. In the test above, nodes 10 and 11 are now
classified as "partly-direct" dependencies because they were introduced
by a direct arc, even though they themselves are ancestral. Pcp stores
such dependencies, which fixes the problem.

(Internal change: 2360903)