Pre-Time support for UsdTimeCode
A UsdTimeCode could also represent a 'pre-time' value.

In a subsequent change this will be used to evaluate attr values at
a pre-time UsdTimeCode. The returned value is the limit when the time
is approached from the left.

Introduces Factory initialization for UsdTimeCode::PreTime.
Updated tests for testUsdTimeCodeRepr, testUsdTimeCode and
testUsdTimeCodeStream

Note that since pre-time is a query only operation, for any pre-time
UsdTimeCode UsdAttribute::Set and UsdAttribute::Clear will result in a
coding error.

(Internal change: 2361353)