Enable querying for preValue for Usd Time Samples evaluation.
The returned value is the value at the limit when the time is approached from
the left.

Note that when querying for a pre-time for an attribute, data sources
could be queried twice, once to get the bracketing time samples and then to
get the previous time sample (this is when querying time lands on an
authored time sample itself).

For a querying time which lands on a time sample itself, a pre-time
evaluation will hold the previous time sample value, unless the values
are linearly interpolatable.

Note that pre-time query support for ValueClips and TsSpline will be added in a
subsequent change.

(Internal change: 2361354)