Fixing a bug where changes to variants (or their dependencies) would cause resyncs and recomposition of prims that do not actually depend on the changed variant. This was due to the dependency code erroneously treating the parent of a variant selection spec (which is prim that owns the variant set spec) as representing an ancestral dependency of the variant path. This is not the case because a variant is not a namespace child of its prim when it comes to composed namespace so we were getting extra ancestral dependencies triggering resyncs when those dependencies do not actually exist.
With this issue fixed though, I had to make an additional change to make
sure that UsdStage still notifies that a change of adding or removing a
variant from a variant set spec causes the composed variants for the
parent prim (as accessed through prim.GetVariantSet(...).GetVariants())
to potentially change. We indicate this by having UsdObjectsChanged add
the prim path to its "changedInfoOnlyPaths" as a change to the
variantChildren field. This allows clients to know if it needs to call
GetVariants or GetVariantNames again for a potential composed variants
change.

(Internal change: 2362029)