usdImagingGL: Fix arguments to testUsdImagingGLTetMesh tests
These tests were missing arguments that were used for the baseline
images. This was not caught earlier because a bug in the idiff
tool from the OpenImageIO version used internally caused it to
incorrectly report that there were no diffs. Switching to OIIO
v2.5.6.0 (from change 2349320) exposed this problem.

(Internal change: 2362293)