usdUI: Accessibility API schemas
This PR adds an Accessibility API schemas to UsdUI as an attempt to
make it a Universally Accessible Scene Description.

A brief overview of the schema:

[Proposal](https://github.com/PixarAnimationStudios/OpenUSD-proposals/tree/main/proposals/accessibility)

```
def "Foo"(
    apiSchemas = ["AccessibilityAPI:default"]
) {
    string accessibility:default:label = "Short Label"
    string accessibility:default:description = "A longer description of this prim"
    uniform token accessibility:default:priority = "standard"
}
```

UsdUIAccessibilityAPI is a multiple apply schema.
Each instance has three attributes:

label: A short description of the prim
description: An extended description of the prim
priority: A hint to a runtime on how this instance should be valued

There are a few small changes from the PR as originally filed:

- Docs in schema.usda have been line-wrapped so that the generated
  code adheres to the 80 character line limit and updated to correct
  a few typos.

- testUsdUIAcessibilityAPI was renamed to correct the typo.

Closes #3271

(Internal change: 2362308)