Autodesk: Fix _HgiToResourceRegistryMap static lifetime destruction bug
If a resource registry has a static lifetime object as its root owner, then we
can encounter a static lifetime ordering issue, since this registry map also
has a static lifetime. It's possible that due to the unspecified ordering of
static destruction, this map is destroyed before a registry, in which case
calling _Unregister() would be undefined behaviour. To prevent this we use a
weak ownership, and skip the call when the map is dead because it no longer
matters.

Closes #3557

(Internal change: 2362312)