Edits through the UsdNamespaceEditor now process the layer changes to provide UsdStage's ObjectChanged notice with more information about what overall edits have occurred to UsdPrims. This can be queried in the sent notice via the function GetPrimResyncType which allows the client to determine if the reported resync paths are effective renames and/or reparents or even composition no-ops due to downstream dependency fixups.
We determine this by having the UsdNamespaceEditor open a special change
block on all its dependent stages with the edits stage level prim moves
it expected to perform. The stage's change process account for these
expected moves and checks that the move results in a prim with the exact
same prim stack as before. If it does, that move is included in the
ObjectsChanged notice the stage sends so that it can be queried via
GetPrimResyncType to allow clients to decide if they can treat the
resync in special (and likely more efficient) manner than they could
without this information.

(Internal change: 2362384)