Fixes a thread-safety issue with concurrently growing the node filter vector.
The concurrent growth methods on tbb::concurrent_vector are thread-safe, with
the exception of invocation of constructors of newly added entries not
synchronizing with concurrent access of those new entries. This can lead to
newly added entries still being under construction, if they are immediately
accessed from another thread, for example, after concurrent calls to
grow_to_at_least() or grow_by() return.

In the past, we have solved this by passing a custom allocator that zero-
initializes memory underpinning tbb::concurrent_vector allocations, and then
relying on the default-initializing constructor of std::atomic not overwriting
that memory at construction time - even though this already dabbles in undefined
behavior. Starting with C++20, the std::atomic constructor will be value-
initializing the underlying integer object, which runs the risk of constructors
resetting the entries to 0 after, for example, atomic increments have already
been performed. We will, however, be able to use std::atomic_ref to replace
this code once C++20 is available to us.

In the meantime, his change attempts to solve the problem in a different way,
using an atomic flag to synchronize on concurrent growth, including constructor
invocation by the first thread that gets to calling grow_to_at_least.

(Internal change: 2362415)