Changes Exec_OutputKeys to store the Esf object along with a pointer to the computation definition.
This makes it harder to inadvertently access the scene through SdfPath objects
without journaling the operations, which can lead to uncompilation and
recompilation not functioning as intended.

This also makes the responsibility of the input resolving task clearer, which is
now consistently on the hook to resolve computation definitions. Previously,
this logic was fractured across the input resolving and output providing tasks.

(Internal change: 2362743)