[hd] Fix HdPrimGather crash when picking in a scene with 80k prims -- due to overflowing the callstack -- by replacing callstack recursion with heap iteration.
In the crashing scenario, there are 80826 include paths,
and 288 exclude paths interleaved semi-uniformly.
The setup is sufficient to make the recursive
algorithm recurse 80k+ levels deep.

We think there are deeper algorithmic improvements possible
here, but this change is intended to provide a low-risk
targetted fix to the current approach.

(Internal change: 2363161)