Implement journaled Provider Resolution
Resolution of computation inputs should implement its scene traversals in
terms of individual hops through the scene graph instead of jumping straight
to the provider with `GetObjectAtPath`. By resolving inputs in this manner,
input resolution populates the EsfJournal with the minimal set of edit
reasons needed to later uncompile that input.

This change introduces a global `Exec_ResolveInput` function that performs
the traversal given an origin object, an input key, and an EsfJournal which
records each hop performed by the traversal. The result of this function is
a set of Exec_OutputKeys which must be compiled and connected to the input
specified by the input key.

This change updates the compilation tasks to use the new `Exec_ResolveInput`
function. The input resolving task invokes the new function with a journal
that's owned by the calling task (either the output providing task, or leaf
compilation task). The calling task uses this journal when connecting their
input to its resolved outputs.

(Internal change: 2363567)