Autodesk: Add the drawingCoordBufferBinding in the Hash computing of ResourceBinder metadata
The DrawingCoordBufferBinding is not included in the hash of
`HdStResourceBinder::MetaData`. So two draw items with the same geometric
shader and different DrawingCoordBufferBinding will use the same culling
program, but they should use different culling programs.

In this PR we add `DrawingCoordBufferBinding` to the hash of
`HdStResourceBinder::MetaData`.

Closes #3585

(Internal change: 2363580)