Deprecate the curve type for TsKnot
Knots still have a curve type but it is ignored and knots can be added
to any spline of the correct data type.

Invoking the TsKnot constructor with a TsCurveType or calling
GetCurveType or SetCurveType will emit a warning message informing the
user that curve types for knots have been deprecated. The knot's curve
type is no longer considered when comparing knots for equality, so
knots with differing curve types but otherwise equal values will
compare equal.

(Internal change: 2364135)