[hd] Fix drawTarget dirtyBits translation
Using HdChangeTracker::DirtyBits to invalidate "everything" on the draw target effectively signaled to HdStDrawTarget that everything except the camera was dirty, because HdChangeTracker::DirtyBits leaves the twos bit (which HdStDrawTarget interprets as DirtyDTCamera) unset. This change fixes that so all bits HdStDrawTarget cares about are set to dirty.

(Internal change: 2364673)