Add execGeom library and schema type hierarchy support.
Exec_DefinitionRegistry now respects the typed schema hierarchy, allowing
computations that are registered on a base type to be found on a prim of a
derived schema type.

Note that the way computation lookup is currently done is inefficient, because
we traverse the schema type hierarchy every time, looking for a shema that
provides the requested computation. We will want to replace this with caching,
probably by storing composed exec prim definitions.

The execGeom library defines an exec computation 'computeTransform' that
computes the local-to-world matrix for a given xformable prim--but only in the
case where all xformable prims in the hierarchy have a single transform
transform op. The intention is not to expose exec behavior that's useful for
clients (yet), but to demonstrate how exec computations will be used as OpenExec
matures.

(Internal change: 2365501)