vdf: fix check for debug name callback validity
Previously, the callable passed to SetDebugNameCallback was captured
in a lambda, which was then wrapped in a VdfNodeDebugNameCallback.
Checking the validity of that callback could never fail because we had
just put a lambda in it.  If we had passed an invalid
VdfNodeDebugNameCallback, we would have smuggled it past the validity
check and then thrown an exception when invoking it later.

Instead, construct a VdfNodeDebugNameCallback directly from the
callable argument.  Also, pass the callback into the network with an
r-value reference and move it into place.

Two internal callback validity checks were changed to be TF_VERIFY.
The check in VdfNode::SetDebugNameCallback is reachable from client
API so it remains a coding error.

(Internal change: 2365960)