exec: fix initialization races in type registry
Like many of our registries that support concurrent lookup,
when two threads raced to query ExecTypeRegistry the first
time, one of them could observe the state of the registry
prior to the completion of registry function subscription.

Refactor the solution from change 2365742 into a reusable component,
Exec_RegistrationBarrier, and use it to fix the issue in
ExecTypeRegistry.

(Internal change: 2365961)