pcp: Use bitmask for excluded tasks in Pcp_PrimIndexer::AddTasksForNode
This function previously used several bool flags to indicate
collections of tasks to exclude. This was inflexible and led to
spaghetti-ish code to resolve the bool flags against each other
to determine the final set of tasks to enqueue. Using a bitmask
avoids these problems. This simplifies an upcoming change that
wants to exclude a different set of tasks in certain situations.

(Internal change: 2367361)