Store node recompilation information in Exec_Program.
Exec_NodeRecompilationInfo objects are created for each node in the network
that supports recompilation. The objects are stored in a concurrent vector
for efficient access. Access to this concurrent vector is managed by an
Exec_NodeRecompilationInfoTable.

The Exec_NodeRecompilationInfo associated with a given node will be
retrieved when exec needs to recompile inputs of that node.

(Internal change: 2367457)