exec: disallow CreateNode from constructing a time input node
Make it clear in the API and the documentation that an ExecSystem has
exactly one time input node, additional time input nodes cannot be
compiled and the existing time input node must not be uncompiled.
Journaled dependencies may not be applied to the time input node.
CreateNode should always create a node.

(Internal change: 2367614)