Fix bug in UsdGeomTetMesh utility function to find inverted elements.
The previous version was flawed and checked the signed volume,
of the parallelpiped, but if all the normals are consistent wrt.
the direction to the center, the volume is always positive
regardless of the orientation attribute value.
Thank you to Simon Schirm of NVidia for flagging this!

The new test verifies that all the face normals point in the correct
direction wrt. the orienation attribute that is set. If any deviate from
this, the element is considered inverted.

Additional test coverage was added to test_UsdGeomTetMeshFindInvertedElements.

(Internal change: 2368092)