github: Update GitHub Actions to use macos-15 and Xcode 16
This switches the runners to use ARM64 Macs instead of older Intel
models. This aligns with the Macs used internally for development.

This change also sidesteps an issue where build_usd.py currently uses
an older version of TBB on Intel Macs that is incompatible with
OpenExec. This will be addressed in a follow-up change.

(Internal change: 2368125)