usd: Fix latent bug in testUsdResolveTarget on Windows
The test previously called TfGetBaseName on anonymous layer
identifiers (e.g. "anon:MEMADDRESS:root-session.usda") expecting
to get back the same identifier for further processing. This
works on Linux and macOS, but returns the empty string on
Windows. I haven't confirmed this, but I think this is because
the ':' characters are drive letter separators, and the presence
of more than one is interpreted as an invalid path.

This is pre-existing behavior that had caused failed verifies
on Windows that had been ignored. However, after change 2368249
the failed verifies caused this test to fail. This change
fixes the test to avoid calling TfGetBaseName in the case
where the layer is expected to be anonymous.

(Internal change: 2368384)