[hgi*, hdSt] Support multiple textures in a single HgiTextureBindDesc in support of arrays of textures.
Change NamedTextureHandle member "handle" to a std::vector of
HdStTextureHandleSharedPtr called "handles." When the size of handles is greater
than 1, treat that NamedTextureHandle as referring to an array of textures.
Currently only the simple lighting shader uses multiple handles within a
NamedTextureHandle, which is done for shadow maps.

This functionality was added because in Vulkan, an array of textures uses one
binding unit, as opposed to GL in which each texture in an array uses its own
binding unit.

(Internal change: 2368478)
(Internal change: 2368498)