usd: Fix expected namespace edit notices in test case
This test case expected to receive an "UnchangedPrimStack" resync
notices from a stage that was _not_ marked as a dependent of the
stage where the namespace edits were invoked. This is incorrect,
since non-dependent stages can only report "Delete" or "Other"
resync notices in response to a namespace edit.

This test is currently commented out but will be enabled by
an upcoming change.

(Internal change: 2368798)