TsSpline::Sample can read uninitialized memory
If sampling a spline whose type is not double and that does not use
inner loops, there is an off-by-one error that fails to copy the last
necessary knot.

We now copy the knot on the far end of the last segment we need to
sample and have added a test for this case.

(Internal change: 2369087)
(Internal change: 2369728)