pcp: Simplified specializes
This change reimplements specializes composition to avoid
the propagation of prim index subtrees to/from the "origin"
specializes node (which represents the location where the
specializes arc was authored) and the root of the prim index
(which represents the strength ordering of the opinions).

I had originally chosen the previous approach because it
kept the specializes behavior self-contained -- other parts
of Pcp didn't need to have specific handling for this arc.
However, the propagation code was complicated and was the
source of a number of correctness bugs. This approach also
duplicated entire subtrees of nodes in prim indexes, although
performance issues may have been largely mitigated by
Pcp's node culling and copy-on-write optimizations.

With this change the propagation scheme and graph structure
are much simpler: when Pcp encounters a specializes arc,
it will add an inert placeholder node to represent the
origin and propagate that node to the root of the prim
index. Any further composition arcs will only be evaluated
on the propagated node and never on the placeholder. This
removes the need to copy subtrees back and forth and
simplifies the remaining code to propagate the placeholder
node. The tradeoff is that a handful of code sites in Pcp
now need to know about the propagated specializes node.
Even with this, the new implementation is still fewer
lines of code than the old one.

This change fixes the correctness bugs documented in the
SepcializesAndVariants museum cases; those baselines have
been updated accordingly. It also enables some namespace
editing functionality that was disabled because of issues
with the duplicated subtrees in the prim index. The
test cases in testUsdNamespaceEditorDependentEditsBasicClassArcs
have been enabled to verify this. The remaining test
updates reflect the removal of the duplicated subtrees.

Fixes #3244

(Internal change: 2369111)