updateSchemaWithSdrNode: Maintain source shader param order
We now forward the parameter order received from SdrNode (which corresponds
to the parsed order in the source shader) through to the propertyOrder
metadata field in the output schema's prim spec.

This also required an update to usdGenSchema which previous to this did not
allow propertyOrder through the schema generation process for API schemas.
After discussion we think it makes sense to allow it for single-apply
schemas only.

An upcoming change will update UsdSchemaRegistry's prim registration logic
to consider API schemas as a source for propertyOrder values. Until then
only Typed schemas will reflect the preserved property ordering this change
provides.

(Internal change: 2369252)